+++ PowerVFX/Lib/PowerVFXPass.cginc
@@ -5,14 +5,14 @@
 
 v2f vert(appdata v)
 {
-    half4 worldPos = mul(unity_ObjectToWorld,v.vertex);
-    half3 viewDir = normalize(UnityWorldSpaceViewDir(worldPos.xyz));
-    half3 worldNormal = normalize(UnityObjectToWorldNormal(v.normal));
+    float4 worldPos = mul(unity_ObjectToWorld,v.vertex);
+    float3 viewDir = normalize(UnityWorldSpaceViewDir(worldPos.xyz));
+    float3 worldNormal = normalize(UnityObjectToWorldNormal(v.normal));
 
     v2f o = (v2f)0;
     o.color = v.color;
     if(_VertexWaveOn){
-        half attemMaskCDATA = v.uv1.z;
+        float attemMaskCDATA = v.uv1.z;
         ApplyVertexWaveWorldSpace(worldPos.xyz/**/,worldNormal,v.color,v.uv,attemMaskCDATA);
     }
     o.vertex = UnityWorldToClipPos(worldPos);
@@ -27,13 +27,13 @@ v2f vert(appdata v)
     //         o.grabPos.y = o.grabPos.w - o.grabPos.y;
     // #endif
 
-    half3 normalDistorted = SafeNormalize(worldNormal + _EnvOffset.xyz);
+    float3 normalDistorted = SafeNormalize(worldNormal + _EnvOffset.xyz);
     if(_EnvReflectOn)
         o.reflectDir = reflect(- viewDir,normalDistorted);
     if(_EnvRefractionOn)
         o.refractDir = refract(viewDir,-normalDistorted,1/_EnvRefractionIOR);
 
-    half3 viewNormal = normalize(mul((half3x3)UNITY_MATRIX_MV,v.normal));
+    float3 viewNormal = normalize(mul((half3x3)UNITY_MATRIX_MV,v.normal));
     o.viewNormal = viewNormal.xy * 0.5 + 0.5;
 
     if(_FresnelOn)
@@ -42,8 +42,8 @@ v2f vert(appdata v)
     o.fresnal_customDataZ.yzw = v.uv1.xyz;// particle custom data (Custom1).zw
 
     if(_LightOn){
-        half3 worldTangent = UnityObjectToWorldDir(v.tangent.xyz);
-        TANGENT_SPACE_COMBINE(worldPos,worldNormal,half4(worldTangent,v.tangent.w),o/**/);
+        float3 worldTangent = UnityObjectToWorldDir(v.tangent.xyz);
+        TANGENT_SPACE_COMBINE(worldPos,worldNormal,float4(worldTangent,v.tangent.w),o/**/);
     }
     UNITY_TRANSFER_FOG(o,o.vertex);
     return o;
@@ -52,29 +52,29 @@ v2f vert(appdata v)
 fixed4 frag(v2f i,fixed faceId:VFACE) : SV_Target
 {
     TANGENT_SPACE_SPLIT(i);
-    half3 reflectDir = i.reflectDir;
-    half3 refractDir = i.refractDir;
+    float3 reflectDir = i.reflectDir;
+    float3 refractDir = i.refractDir;
 
-    half4 mainColor = half4(0,0,0,1);
-    half4 screenColor=0;
+    float4 mainColor = float4(0,0,0,1);
+    float4 screenColor=0;
     // setup mainUV, move to vs
-    // half4 mainUV = MainTexOffset(i.uv);
-    half4 mainUV = i.uv;
+    // float4 mainUV = MainTexOffset(i.uv);
+    float4 mainUV = i.uv;
 
 // get particle system's custom data
-    half dissolveCustomData = i.fresnal_customDataZ.y;
-    half dissolveEdgeWidth = i.fresnal_customDataZ.z;
-    half distortionCustomData = i.fresnal_customDataZ.w;
+    float dissolveCustomData = i.fresnal_customDataZ.y;
+    float dissolveEdgeWidth = i.fresnal_customDataZ.z;
+    float distortionCustomData = i.fresnal_customDataZ.w;
 
     //use _CameraOpaqueTexture
-    half2 screenUV = i.grabPos.xy/i.grabPos.w;
+    float2 screenUV = i.grabPos.xy/i.grabPos.w;
     mainUV.xy = _MainTexUseScreenColor == 0 ? mainUV.xy : screenUV;
     
-    half2 uvDistorted = mainUV.zw;
+    float2 uvDistorted = mainUV.zw;
     if(_DistortionOn){
-        half4 distortUV = mainUV.zwzw * _DistortTile + frac(_DistortDir * _Time.xxxx);
+        float4 distortUV = mainUV.zwzw * _DistortTile + frac(_DistortDir * _Time.xxxx);
         if(_DistortionRadialUVOn){
-            half4 p = _DistortionRadialCenter_LenScale_LenOffset;
+            float4 p = _DistortionRadialCenter_LenScale_LenOffset;
             distortUV.xy = PolarUV(mainUV.zw,p.xy,p.z,p.w*_Time.x,_DistortionRadialRot);
         }
         uvDistorted = ApplyDistortion(mainUV,distortUV,distortionCustomData);
@@ -91,26 +91,26 @@ fixed4 frag(v2f i,fixed faceId:VFACE) : SV_Target
         ApplyEnv(mainColor,mainUV.zw,reflectDir,refractDir);
 
     if(_OffsetOn){
-        half4 offsetUV = (_ApplyToOffset ? uvDistorted.xyxy : mainUV.zwzw) * _OffsetTile + (_Time.xxxx * _OffsetDir); //暂时去除 frac
+        float4 offsetUV = (_ApplyToOffset ? uvDistorted.xyxy : mainUV.zwzw) * _OffsetTile + (_Time.xxxx * _OffsetDir); //暂时去除 frac
         if(_OffsetRadialUVOn){
-            half4 p = _OffsetRadialCenter_LenScale_LenOffset;
+            float4 p = _OffsetRadialCenter_LenScale_LenOffset;
             offsetUV.xy = PolarUV(mainUV.zw,p.xy,p.z,p.w*_Time.x,_OffsetRadialRot);
         }
-        // half2 maskUVOffset = _OffsetMaskTex_ST.zw * (1 + _Time.xx *(1- _OffsetMaskPanStop) );
-        half2 maskUVOffset = UVOffset(_OffsetMaskTex_ST.zw, _OffsetMaskPanStop);
-        half2 maskUV = mainUV.zw * _OffsetMaskTex_ST.xy + maskUVOffset;
+        // float2 maskUVOffset = _OffsetMaskTex_ST.zw * (1 + _Time.xx *(1- _OffsetMaskPanStop) );
+        float2 maskUVOffset = UVOffset(_OffsetMaskTex_ST.zw, _OffsetMaskPanStop);
+        float2 maskUV = mainUV.zw * _OffsetMaskTex_ST.xy + maskUVOffset;
         ApplyOffset(mainColor,offsetUV,maskUV);
     }
 
     //dissolve
     if(_DissolveOn){
-        half2 dissolveUVOffset = UVOffset(_DissolveTex_ST.zw,_DissolveTexOffsetStop);
-        half2 dissolveUV = mainUV.zw * _DissolveTex_ST.xy + dissolveUVOffset;
+        float2 dissolveUVOffset = UVOffset(_DissolveTex_ST.zw,_DissolveTexOffsetStop);
+        float2 dissolveUV = mainUV.zw * _DissolveTex_ST.xy + dissolveUVOffset;
         ApplyDissolve(mainColor,dissolveUV,i.color,dissolveCustomData,dissolveEdgeWidth);
     }
 
     if(_FresnelOn){
-        half fresnal = i.fresnal_customDataZ.x;
+        float fresnal = i.fresnal_customDataZ.x;
         ApplyFresnal(mainColor,fresnal,screenColor);
     }
